pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-drive"
version = "0.0.1"
description = "Encrypted drive with Streamlit UI"
requires-python = ">=3.11"
dependencies = [
  "cryptography==42.0.8",
  "argon2-cffi==23.1.0",
  "pydantic==2.9.2",
  "streamlit==1.38.0",
  "sqlalchemy==2.0.36",
  "boto3==1.34.162",
  "python-dotenv==1.0.1",
  "pytest==8.3.2",
  "pytest-cov==5.0.0",
  "mypy==1.11.2",
  "ruff==0.6.8",
]

# Extra opcional para Postgres (compatible con Python 3.13).
[project.optional-dependencies]
postgres = ["psycopg[binary]==3.1.19"]

[tool.setuptools.packages.find]
include = ["core*", "api*"]
exclude = ["app_streamlit*", "infra*", "tests*"]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP"]
exclude = ["infra", ".venv", ".git"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true
